Add files via upload
A social media dashboard created in Power BI that allows the user to search for a South African social media creator and check relevant indicators such as the number of followers, views, likes, comments and the revenue they generate. The dataset used was generated using a ChatGPT prompt and is not a reflection of real life. The data consists of the username of the creator, the platforms they create on, the region and state they live in.